Distribute values in fact tables to handle long lasting sessions, updating content length field in HTTP related operation from int to long
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@37143 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ http-casing/src/com/untangle/node/http/HttpParser.java
@@ -74,8 +74,8 @@ public class HttpParser extends AbstractParser
 
     private int state;
     private int transferEncoding;
-    private int contentLength; /* counts down content-length and chunks */
-    private int lengthCounter; /* counts up to final */
+    private long contentLength; /* counts down content-length and chunks */
+    private long lengthCounter; /* counts up to final */
 
     // constructors -----------------------------------------------------------
 
@@ -754,7 +754,7 @@ public class HttpParser extends AbstractParser
                 logger.debug(sessStr + "using content length encoding");
             }
             transferEncoding = CONTENT_LENGTH_ENCODING;
-            contentLength = Integer.parseInt(value);
+            contentLength = Long.parseLong(value);
             if (logger.isDebugEnabled()) {
                 logger.debug(sessStr + "CL contentLength = " + contentLength);
             }
